+++ RELEASES.md
@@ -7,6 +7,7 @@ Upcoming
 --------
 
 - Fix text color and errors with actions after Run, Replace in Test tab
+- Improve Open after Run (select report and don't compare and sort all reports in Compare tab)